Create  CloudWatch_alarm_triggers.py
You are a DevOps engineer working for a company that hosts a popular website. The website is hosted on an EC2 instance in Amazon EC2. You need to create a CloudWatch alarm that triggers an alert when the CPU usage of the EC2 instance exceeds 80% for five consecutive minutes. Write a Python script for this.